Remove unnecessary padding hook
