+++ packages/victory-core/src/victory-util/hooks/index.ts
@@ -3,5 +3,4 @@ export { useAnimationState } from "./use-animation-state";
 export { useDomain } from "./use-domain";
 export { useScale } from "./use-scale";
 export { useRange } from "./use-range";
-export { usePadding } from "./use-padding";
 export { useData } from "./use-data";
